Update yamllint configuration (#3)
Update the configuration of yamllint's `truthy` rule to NOT check keys
for truthy values. Enforcing this rule in this way was incorrect, keys
are not values and thus cannot be truthy either. Moreover, the key "on"
should never be disallowed.

This changes allows for the removal of inline rule ignores in GitHub
Actions workflow files.