Track tooling versions (#9)
* Add `.tool-versions` file

Create the `.tool-versions` file as a means of recording the versions of
tools used by this project. Additionally, asdf [1] can be used to
install these tools in a convenient manner. Of the tooling used by this
project (markdownlint, shellcheck, and yamllint) only those with asdf
plugins are listed.

* Use `.tool-versions` in continuous integration

Update continuous integration workflows that use tooling listed in the
`.tool-versions` to install the respective tooling using the version
listed there. This aims to improve reproducibility.

--
1. https://asdf-vm.com/