Update "Publish / GitHub Release" job
Update the CI job that creates a GitHub Release to include an explicit
`body` (release text) in order to avoid GitHub displaying the commit
message of the commit associated with the tag that the release belongs
to.

The text is set to something simple just to avoid that, given the
simplicity of this project I do not currently find it necessary to go
through the trouble of setting something up to extract the relevant
changes from `CHANGELOG.md`.