Stop using markdownlint
Remove the use of markdownlint and corresponding Make command and
continuous integration job.

The motivation for removing it is that it's installation is non-ideal (a
global npm install) and not automated nor document. Also, the version is
not recorded in any way.

I'd be in favor of re-adding it (or adding a similar tool) if the above
concerns can be addressed. For now though, the MarkDown is edited
infrequently enough to consider removing it a worth-it trade-off.