Drop dev dependency on `act` (#65)
Using `act` has not proven to be particularly useful while adding a
maintenance burden in terms of keeping it up-to-date. Hence, this
removes the `act` configuration from the project. Of course, it's still
possible for contributors to use `act` if they so choose.

See [1] for how `act` was initially added.

--
1. d5c95b522325a6ea2c81c35b1ac60dd33d8d8de7