Use token to create tooling update Pull Requests
Add the use of `tibdex/github-app-token` to the tooling job to generate
a token that can be used to create a Pull Request for which CI jobs are
run. This is an improvement over the status quo as it will remove the
need to either 1) create a manual commit to trigger the CI or 2) ignore
the absence of CI checks.