Refactor to extract helpers and assign proper license
Refactor the code base in order to extract the helper functions related
to GitHub Actions which I took from another project of mine which is
permissively licensed. Hence, to preserve that license the code is now
in it's own file (properly vendored) where it can clearly use that
license. This simultaneously updates the helper functions to more recent
versions which fixes some bugs (related to multiline logging), which are
useful in the long run but don't affect the project in its current state